efi_loader: Add check for event log passed from firmware
Platforms may have support to measure their initial firmware components
and pass the event log to u-boot. The event log address can be passed
in property tpm_event_log_addr and tpm_event_log_size of the tpm node.
Platforms may choose their own specific mechanism to do so. A weak
function is added to check if even log has been passed to u-boot
from earlier firmware components. If available, further event log is
appended to the passed log.

The functionality needs to be further extended to do the following
- Parse and check if event log received from earlier firmware components
is correct
- The earlier firmware components may not have a TPM driver to extend
the PCR's. Though this is not an ideal situation. However for platforms
not supporting it, we can add support to parse and extend the PCR's from
u-boot for eventlog passed from previous firmwares.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>