efi_loader: Add option to extend PCR's for firmware measurements
Firmwares before U-Boot may be capable of doing tpm measurements
and passing them to U-Boot in the form of eventlog. However there
may be scenarios where the firmwares don't have TPM driver and
are not capable of extending the measurements in the PCRs.
the PCR's.

Though this is not an ideal situation. However for platforms not
supporting it, this CONFIG option can be used extend the PCR's from
u-boot for eventlog passed from previous firmwares.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>